+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 - removed the examples dir
 - removed the scripts dir
 - removed mermaid2 support from the docs
+- added encoding types to all open commands
 
 ### Fixes